+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 * Removed deletion on map.
 * Rename `view` to `sequence` and `view-single` to `single`.
+* `in` does nothing if the path does not exist (unlike `update-in`).
 
 0.2.0
 -----